docs: move tutorial to another section (#36)
This branch moves the connector tutorial to another section. Meaning that on https://docs.datalbry.io/ would be another section called Tutorials which contains the said tutorial.

Additionally the links in the README.md have been fixed (not existing markdown files). Now they point instead to the templates in templates.